moving table rendering to commons
